refactor: composable for creating event
